Rename piece -> pieceSvg
