+++ components/PieceSvg.tsx
